Added Identity matrix, changed Upper/LowerTriange and ThreadPoolParams interface

+++ src/main/java/com/lukastack/projectmatrix/profiler/ProfilerTimeUnit.java
@@ -2,10 +2,10 @@ package com.lukastack.projectmatrix.profiler;
 
 public enum ProfilerTimeUnit {
     NANOSECONDS("ns", nanoTime -> nanoTime),
-    MICROSECONDS("µsec", nanoTime -> (long) (nanoTime / Math.pow(10, 4))),
-    MILLISECONDS("ms", nanoTime -> (long) (nanoTime / Math.pow(10, 6))),
-    SECONDS("s", nanoTime -> (long) (nanoTime / Math.pow(10, 9))),
-    MINUTES("mins", nanoTime -> (long) (nanoTime / (6 * Math.pow(10, 10))));
+    MICROSECONDS("µsec", nanoTime -> (nanoTime / Math.pow(10, 4))),
+    MILLISECONDS("ms", nanoTime -> (nanoTime / Math.pow(10, 6))),
+    SECONDS("s", nanoTime -> (nanoTime / Math.pow(10, 9))),
+    MINUTES("mins", nanoTime -> (nanoTime / (6 * Math.pow(10, 10))));
 
     private final String name;
     private final ProfilerUnitConverter converter;
@@ -16,7 +16,7 @@ public enum ProfilerTimeUnit {
         this.converter = converter;
     }
 
-    public long getTime(long nanoTime) {
+    public double getTime(long nanoTime) {
 
         return converter.convert(nanoTime);
     }
