+++ src/main/java/com/lukastack/projectmatrix/profiler/ProfilerUnitConverter.java
@@ -3,5 +3,5 @@ package com.lukastack.projectmatrix.profiler;
 @FunctionalInterface
 interface ProfilerUnitConverter {
 
-    long convert(long time);
+    double convert(long time);
 }
